finished lab8 program 1

+++ labs/lab08/r_structs.h
@@ -12,5 +12,5 @@ struct UndergradStudents {
     string firstName;
     string lastName;
     string major;
-    float GPA;
+    double GPA;
 };
